reorganize polygon package
