--- rest/reference/market_test.go
@@ -1,93 +1,0 @@
-package reference_test
-
-import (
-	"context"
-	"encoding/json"
-	"testing"
-
-	"github.com/jarcoal/httpmock"
-	polygon "github.com/polygon-io/client-go"
-	"github.com/stretchr/testify/assert"
-)
-
-func TestGetMarketHolidays(t *testing.T) {
-	c := polygon.NewClient("API_KEY")
-
-	httpmock.ActivateNonDefault(c.Reference.HTTP.GetClient())
-	defer httpmock.DeactivateAndReset()
-
-	expectedResponse := `[
-	{
-		"exchange": "NYSE",
-		"name": "Thanksgiving",
-		"date": "2020-11-26",
-		"status": "closed"
-	},
-	{
-		"exchange": "NASDAQ",
-		"name": "Thanksgiving",
-		"date": "2020-11-26",
-		"status": "closed"
-	},
-	{
-		"exchange": "OTC",
-		"name": "Thanksgiving",
-		"date": "2020-11-26",
-		"status": "closed"
-	},
-	{
-		"exchange": "NASDAQ",
-		"name": "Thanksgiving",
-		"date": "2020-11-27",
-		"status": "early-close",
-		"open": "2020-11-27T14:30:00.000Z",
-		"close": "2020-11-27T18:00:00.000Z"
-	},
-	{
-		"exchange": "NYSE",
-		"name": "Thanksgiving",
-		"date": "2020-11-27",
-		"status": "early-close",
-		"open": "2020-11-27T14:30:00.000Z",
-		"close": "2020-11-27T18:00:00.000Z"
-	}
-]`
-
-	registerResponder("https://api.polygon.io/v1/marketstatus/upcoming", expectedResponse)
-	res, err := c.Reference.GetMarketHolidays(context.Background())
-
-	assert.Nil(t, err)
-	b, err := json.MarshalIndent(res, "", "\t")
-	assert.Nil(t, err)
-	assert.Equal(t, expectedResponse, string(b))
-}
-
-func TestGetMarketStatus(t *testing.T) {
-	c := polygon.NewClient("API_KEY")
-
-	httpmock.ActivateNonDefault(c.Reference.HTTP.GetClient())
-	defer httpmock.DeactivateAndReset()
-
-	expectedResponse := `{
-	"afterHours": true,
-	"currencies": {
-		"crypto": "open",
-		"fx": "open"
-	},
-	"earlyHours": false,
-	"exchanges": {
-		"nasdaq": "extended-hours",
-		"nyse": "extended-hours",
-		"otc": "closed"
-	},
-	"market": "extended-hours",
-	"serverTime": "2020-11-10T22:37:37.000Z"
-}`
-
-	registerResponder("https://api.polygon.io/v1/marketstatus/now", expectedResponse)
-	res, err := c.Reference.GetMarketStatus(context.Background())
-	assert.Nil(t, err)
-	b, err := json.MarshalIndent(res, "", "\t")
-	assert.Nil(t, err)
-	assert.Equal(t, expectedResponse, string(b))
-}
