--- rest/reference/reference.go
@@ -1,8 +1,0 @@
-package reference
-
-import "github.com/polygon-io/client-go/rest/client"
-
-// Client defines a REST client for the Polygon reference API.
-type Client struct {
-	client.Client
-}
