+++ rest/snapshot.go
@@ -1,4 +1,4 @@
-package snapshot
+package polygon
 
 import (
 	"context"
@@ -8,14 +8,14 @@ import (
 	"github.com/polygon-io/client-go/rest/models"
 )
 
-// Client defines a REST client for the Polygon snapshot API.
-type Client struct {
+// SnapshotClient defines a REST client for the Polygon snapshot API.
+type SnapshotClient struct {
 	client.Client
 }
 
 // GetAllTickersSnapshot gets the current minute, day, and previous day’s aggregate, as well as the last trade and quote for all symbols of a specified market type.
 // For more details see https://polygon.io/docs/stocks/get_v2_snapshot_locale_us_markets_stocks_tickers.
-func (ac *Client) GetAllTickersSnapshot(ctx context.Context, params *models.GetAllTickersSnapshotParams, opts ...models.RequestOption) (*models.GetAllTickersSnapshotResponse, error) {
+func (ac *SnapshotClient) GetAllTickersSnapshot(ctx context.Context, params *models.GetAllTickersSnapshotParams, opts ...models.RequestOption) (*models.GetAllTickersSnapshotResponse, error) {
 	res := &models.GetAllTickersSnapshotResponse{}
 	err := ac.Call(ctx, http.MethodGet, models.GetAllTickersSnapshotPath, params, res, opts...)
 	return res, err
@@ -23,7 +23,7 @@ func (ac *Client) GetAllTickersSnapshot(ctx context.Context, params *models.GetA
 
 // GetTickerSnapshot gets the current minute, day, and previous day’s aggregate, as well as the last trade and quote for a single traded symbol of a specified market type.
 // For more details see https://polygon.io/docs/stocks/get_v2_snapshot_locale_us_markets_stocks_tickers__stocksticker.
-func (ac *Client) GetTickerSnapshot(ctx context.Context, params *models.GetTickerSnapshotParams, opts ...models.RequestOption) (*models.GetTickerSnapshotResponse, error) {
+func (ac *SnapshotClient) GetTickerSnapshot(ctx context.Context, params *models.GetTickerSnapshotParams, opts ...models.RequestOption) (*models.GetTickerSnapshotResponse, error) {
 	res := &models.GetTickerSnapshotResponse{}
 	err := ac.Call(ctx, http.MethodGet, models.GetTickerSnapshotPath, params, res, opts...)
 	return res, err
@@ -31,7 +31,7 @@ func (ac *Client) GetTickerSnapshot(ctx context.Context, params *models.GetTicke
 
 // GetGainersLosersSnapshot gets the current top 20 gainers or losers of the day in a specific market type.
 // For more details see https://polygon.io/docs/stocks/get_v2_snapshot_locale_us_markets_stocks__direction.
-func (ac *Client) GetGainersLosersSnapshot(ctx context.Context, params *models.GetGainersLosersSnapshotParams, opts ...models.RequestOption) (*models.GetGainersLosersSnapshotResponse, error) {
+func (ac *SnapshotClient) GetGainersLosersSnapshot(ctx context.Context, params *models.GetGainersLosersSnapshotParams, opts ...models.RequestOption) (*models.GetGainersLosersSnapshotResponse, error) {
 	res := &models.GetGainersLosersSnapshotResponse{}
 	err := ac.Call(ctx, http.MethodGet, models.GetGainersLosersSnapshotPath, params, res, opts...)
 	return res, err
@@ -39,7 +39,7 @@ func (ac *Client) GetGainersLosersSnapshot(ctx context.Context, params *models.G
 
 // GetOptionContractSnapshot gets the snapshot of an option contract for a stock equity.
 // For more details see https://polygon.io/docs/options/get_v3_snapshot_options__underlyingasset___optioncontract.
-func (ac *Client) GetOptionContractSnapshot(ctx context.Context, params *models.GetOptionContractSnapshotParams, opts ...models.RequestOption) (*models.GetOptionContractSnapshotResponse, error) {
+func (ac *SnapshotClient) GetOptionContractSnapshot(ctx context.Context, params *models.GetOptionContractSnapshotParams, opts ...models.RequestOption) (*models.GetOptionContractSnapshotResponse, error) {
 	res := &models.GetOptionContractSnapshotResponse{}
 	err := ac.Call(ctx, http.MethodGet, models.GetOptionContractSnapshotPath, params, res, opts...)
 	return res, err
@@ -47,7 +47,7 @@ func (ac *Client) GetOptionContractSnapshot(ctx context.Context, params *models.
 
 // GetCryptoFullBookSnapshot gets the current level 2 book of a single cryptocurrency ticker. This is the combined book from all of the exchanges.
 // For more details see https://polygon.io/docs/crypto/get_v2_snapshot_locale_global_markets_crypto_tickers__ticker__book.
-func (ac *Client) GetCryptoFullBookSnapshot(ctx context.Context, params *models.GetCryptoFullBookSnapshotParams, opts ...models.RequestOption) (*models.GetCryptoFullBookSnapshotResponse, error) {
+func (ac *SnapshotClient) GetCryptoFullBookSnapshot(ctx context.Context, params *models.GetCryptoFullBookSnapshotParams, opts ...models.RequestOption) (*models.GetCryptoFullBookSnapshotResponse, error) {
 	res := &models.GetCryptoFullBookSnapshotResponse{}
 	err := ac.Call(ctx, http.MethodGet, models.GetCryptoFullBookSnapshotPath, params, res, opts...)
 	return res, err
