+++ rest/snapshot_test.go
@@ -1,14 +1,12 @@
-package snapshot_test
+package polygon_test
 
 import (
 	"context"
 	"encoding/json"
-	"net/http"
-	"strings"
 	"testing"
 
 	"github.com/jarcoal/httpmock"
-	polygon "github.com/polygon-io/client-go"
+	polygon "github.com/polygon-io/client-go/rest"
 	"github.com/polygon-io/client-go/rest/models"
 	"github.com/stretchr/testify/assert"
 )
@@ -110,7 +108,7 @@ var snapshot2 = `{
 }`
 
 func TestListSnapshotAllTickers(t *testing.T) {
-	c := polygon.NewClient("API_KEY")
+	c := polygon.New("API_KEY")
 
 	httpmock.ActivateNonDefault(c.Aggs.HTTP.GetClient())
 	defer httpmock.DeactivateAndReset()
@@ -137,7 +135,7 @@ func TestListSnapshotAllTickers(t *testing.T) {
 }
 
 func TestGetTickerSnapshot(t *testing.T) {
-	c := polygon.NewClient("API_KEY")
+	c := polygon.New("API_KEY")
 
 	httpmock.ActivateNonDefault(c.Aggs.HTTP.GetClient())
 	defer httpmock.DeactivateAndReset()
@@ -162,7 +160,7 @@ func TestGetTickerSnapshot(t *testing.T) {
 }
 
 func TestGetGainersLosersSnapshot(t *testing.T) {
-	c := polygon.NewClient("API_KEY")
+	c := polygon.New("API_KEY")
 
 	httpmock.ActivateNonDefault(c.Aggs.HTTP.GetClient())
 	defer httpmock.DeactivateAndReset()
@@ -190,7 +188,7 @@ func TestGetGainersLosersSnapshot(t *testing.T) {
 }
 
 func TestGetOptionContractSnapshot(t *testing.T) {
-	c := polygon.NewClient("API_KEY")
+	c := polygon.New("API_KEY")
 
 	httpmock.ActivateNonDefault(c.Aggs.HTTP.GetClient())
 	defer httpmock.DeactivateAndReset()
@@ -260,7 +258,7 @@ func TestGetOptionContractSnapshot(t *testing.T) {
 }
 
 func TestGetCryptoFullBookSnapshot(t *testing.T) {
-	c := polygon.NewClient("API_KEY")
+	c := polygon.New("API_KEY")
 
 	httpmock.ActivateNonDefault(c.Aggs.HTTP.GetClient())
 	defer httpmock.DeactivateAndReset()
@@ -314,25 +312,4 @@ func TestGetCryptoFullBookSnapshot(t *testing.T) {
 	b, err := json.MarshalIndent(res, "", "\t")
 	assert.Nil(t, err)
 	assert.Equal(t, expectedResponse, string(b))
-}
-
-func registerResponder(url, body string) {
-	httpmock.RegisterResponder("GET", url,
-		func(req *http.Request) (*http.Response, error) {
-			resp := httpmock.NewStringResponse(200, body)
-			resp.Header.Add("Content-Type", "application/json")
-			return resp, nil
-		},
-	)
-}
-
-func indent(first bool, data, indent string) string {
-	lines := strings.Split(data, "\n")
-	for i := range lines {
-		if i == 0 && !first {
-			continue
-		}
-		lines[i] = indent + lines[i]
-	}
-	return strings.Join(lines, "\n")
 }
