+++ rest/trades.go
@@ -1,4 +1,4 @@
-package trades
+package polygon
 
 import (
 	"context"
@@ -9,8 +9,8 @@ import (
 	"github.com/polygon-io/client-go/rest/models"
 )
 
-// Client defines a REST client for the Polygon trades API.
-type Client struct {
+// TradesClient defines a REST client for the Polygon trades API.
+type TradesClient struct {
 	client.Client
 }
 
@@ -38,7 +38,7 @@ func (it *ListTradesIter) Trade() models.Trade {
 //   if iter.Err() != nil {
 //       return err
 //   }
-func (c *Client) ListTrades(ctx context.Context, params *models.ListTradesParams, options ...models.RequestOption) *ListTradesIter {
+func (c *TradesClient) ListTrades(ctx context.Context, params *models.ListTradesParams, options ...models.RequestOption) *ListTradesIter {
 	return &ListTradesIter{
 		Iter: iter.NewIter(ctx, models.ListTradesPath, params, func(uri string) (iter.ListResponse, []interface{}, error) {
 			res := &models.ListTradesResponse{}
@@ -56,7 +56,7 @@ func (c *Client) ListTrades(ctx context.Context, params *models.ListTradesParams
 
 // GetLastTrade retrieves the last trade for a specified ticker.
 // For more details see https://polygon.io/docs/stocks/get_v2_last_trade__stocksticker.
-func (c *Client) GetLastTrade(ctx context.Context, params *models.GetLastTradeParams, options ...models.RequestOption) (*models.GetLastTradeResponse, error) {
+func (c *TradesClient) GetLastTrade(ctx context.Context, params *models.GetLastTradeParams, options ...models.RequestOption) (*models.GetLastTradeResponse, error) {
 	res := &models.GetLastTradeResponse{}
 	err := c.Call(ctx, http.MethodGet, models.GetLastTradePath, params, res, options...)
 	return res, err
@@ -64,7 +64,7 @@ func (c *Client) GetLastTrade(ctx context.Context, params *models.GetLastTradePa
 
 // GetLastCryptoTrade retrieves the last trade for a crypto pair.
 // For more details see https://polygon.io/docs/crypto/get_v1_last_crypto__from___to.
-func (c *Client) GetLastCryptoTrade(ctx context.Context, params *models.GetLastCryptoTradeParams, options ...models.RequestOption) (*models.GetLastCryptoTradeResponse, error) {
+func (c *TradesClient) GetLastCryptoTrade(ctx context.Context, params *models.GetLastCryptoTradeParams, options ...models.RequestOption) (*models.GetLastCryptoTradeResponse, error) {
 	res := &models.GetLastCryptoTradeResponse{}
 	err := c.Call(ctx, http.MethodGet, models.GetLastCryptoTradePath, params, res, options...)
 	return res, err
